visibility radius added

+++ src/header.cpp
@@ -13,7 +13,7 @@ bool operator==(const Coord& c1, const Coord& c2) { return c1.x == c2.x && c1.y 
 Coord operator+(const Coord& left, const Coord& right) { return Coord(left.x + right.x, left.y + right.y); }
 Coord operator-(const Coord& left, const Coord& right) { return Coord(left.x - right.x, left.y - right.y); }
 // Coord Coord::operator/(float del) { return Coord(x / del, y / del); }
-Coord Coord::del(float d) const { return Coord(x / d, y / d); }
+Coord Coord::del(float d) const { return Coord(static_cast<int>(x / d), static_cast<int>(y / d)); }
 
 std::string Coord::toString() const { return std::string("[") + std::to_string(x) + "," + std::to_string(y) + "]"; }
 Distance Coord::length() const { return sqrt(x*x + y*y); }
